refacto: create a useLocationModal hook to handle the shared logic
