+++ src/features/location/components/SearchLocationModal.tsx
@@ -1,14 +1,12 @@
-import React, { useEffect, useState, useCallback } from 'react'
+import React, { useState, useCallback } from 'react'
 import styled from 'styled-components/native'
 
 import { LocationModalButton } from 'features/location/components/LocationModalButton'
 import { LOCATION_PLACEHOLDER } from 'features/location/constants'
 import { LocationMode } from 'features/location/enums'
 import { isCurrentLocationMode } from 'features/location/helpers/locationHelpers'
-import { useCustomLocationModeColor } from 'features/location/helpers/useCustomLocationModeColor'
-import { useGeolocationModeColor } from 'features/location/helpers/useGeolocationModeColor'
-import { GeolocPermissionState, useLocation } from 'libs/geolocation'
-import { SuggestedPlace } from 'libs/place'
+import { useLocationModal } from 'features/location/helpers/useLocationModal'
+import { GeolocPermissionState } from 'libs/geolocation'
 import { LocationSearchFilters } from 'shared/location/LocationSearchFilters'
 import { LocationSearchInput } from 'shared/location/LocationSearchInput'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
@@ -35,46 +33,25 @@ export const SearchLocationModal = ({
 }: LocationModalProps) => {
   const {
     isGeolocated,
-    isCustomPosition,
-    place,
-    setPlace: setPlaceGlobally,
+    placeQuery,
+    setPlaceQuery,
+    selectedPlace,
+    setSelectedPlace,
+    selectedLocationMode,
+    setSelectedLocationMode,
+    geolocationModeColor,
+    customLocationModeColor,
+    onSetSelectedPlace,
+    onResetPlace,
+    setPlaceGlobally,
     onModalHideRef,
     permissionState,
     requestGeolocPermission,
     showGeolocPermissionModal,
-  } = useLocation()
+  } = useLocationModal(visible)
 
-  const [placeQuery, setPlaceQuery] = useState('')
   const [aroundRadius, setAroundRadius] = useState(100)
   const [includeDigitalOffers, setIncludeDigitalOffers] = useState(false)
-  const [selectedPlace, setSelectedPlace] = useState<SuggestedPlace | null>(null)
-  const defaultLocationMode = isGeolocated ? LocationMode.GEOLOCATION : LocationMode.NONE
-  const [selectedLocationMode, setSelectedLocationMode] =
-    useState<LocationMode>(defaultLocationMode)
-
-  const geolocationModeColor = useGeolocationModeColor(selectedLocationMode)
-  const customLocationModeColor = useCustomLocationModeColor(selectedLocationMode)
-
-  const initializeLocationMode = useCallback(() => {
-    onModalHideRef.current = undefined
-    if (isCustomPosition) {
-      setSelectedLocationMode(LocationMode.CUSTOM_POSITION)
-    } else {
-      setSelectedLocationMode(defaultLocationMode)
-    }
-  }, [onModalHideRef, isCustomPosition, setSelectedLocationMode, defaultLocationMode])
-
-  useEffect(() => {
-    if (visible) {
-      initializeLocationMode()
-      if (place) {
-        onSetSelectedPlace(place)
-      } else {
-        onResetPlace()
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [visible, initializeLocationMode])
 
   const runGeolocationDialogs = useCallback(async () => {
     const selectGeoLocationMode = () => setSelectedLocationMode(LocationMode.GEOLOCATION)
@@ -94,6 +71,7 @@ export const SearchLocationModal = ({
     onModalHideRef,
     showGeolocPermissionModal,
     requestGeolocPermission,
+    setSelectedLocationMode,
   ])
 
   const selectLocationMode = useCallback(
@@ -106,16 +84,6 @@ export const SearchLocationModal = ({
     },
     [runGeolocationDialogs, setSelectedLocationMode]
   )
-
-  const onResetPlace = () => {
-    setSelectedPlace(null)
-    setPlaceQuery('')
-  }
-
-  const onSetSelectedPlace = (place: SuggestedPlace) => {
-    setSelectedPlace(place)
-    setPlaceQuery(place.label)
-  }
 
   const onSubmit = () => {
     if (selectedLocationMode === LocationMode.CUSTOM_POSITION) {
